[FIX] Auth mutation
- Add AuthMutation to RootMutation
- Move login, logout and refresh token from AuthQuery to AuthMutation
- Modify permissions field in UserType, LoginUserResponseType
