+++ timetracker.Server/API/Auth/Types/LoginUserResponseType.cs
@@ -11,7 +11,7 @@ namespace timetracker.Server.API.Auth.Types
             Field(t => t.Name);
             Field(t => t.EmployeeType);
             Field<ListGraphType<StringGraphType>>("permissions")
-                 .Resolve(context => context.Source.Permissions.Split(',').ToList());
+                 .Resolve(context => context.Source.Permissions.Split(',', StringSplitOptions.RemoveEmptyEntries));
         }
     }
 }
