+++ timetracker.Server/API/RootMutation.cs
@@ -1,4 +1,5 @@
 ﻿using GraphQL.Types;
+using timetracker.Server.API.Auth;
 using timetracker.Server.API.User;
 
 namespace timetracker.Server.API
@@ -8,7 +9,7 @@ namespace timetracker.Server.API
         public RootMutation()
         {
             Field<UserMutation>("users").Resolve(_ => new { });
-
+            Field<AuthMutation>("auth").Resolve(_ => new { });
         }
     }
 }
